IGRFBuilderWithLocationTests
